Fix circular dependency and type
