+++ ee/tasks/webhooks_ee.py
@@ -9,9 +9,9 @@ from statshog.defaults.django import statsd
 
 from ee.clickhouse.models.element import chain_to_elements
 from posthog.celery import app
+from posthog.constants import AvailableFeature
 from posthog.models import Action, Event, Team
 from posthog.tasks.webhooks import determine_webhook_type, get_formatted_message
-from posthog.types import AvailableFeature
 
 
 @app.task(ignore_result=True, bind=True, max_retries=3)
