+++ posthog/api/organization.py
@@ -8,6 +8,7 @@ from rest_framework.request import Request
 
 from posthog.api.routing import StructuredViewSetMixin
 from posthog.api.shared import TeamBasicSerializer
+from posthog.constants import AvailableFeature
 from posthog.event_usage import report_onboarding_completed
 from posthog.mixins import AnalyticsDestroyModelMixin
 from posthog.models import Organization, User
@@ -18,7 +19,6 @@ from posthog.permissions import (
     OrganizationMemberPermissions,
     extract_organization,
 )
-from posthog.types import AvailableFeature
 
 
 class PremiumMultiorganizationPermissions(permissions.BasePermission):
