+++ posthog/types.py
@@ -1,4 +1,3 @@
-from enum import Enum
 from typing import Union
 
 from posthog.models.filters.filter import Filter
@@ -7,12 +6,3 @@ from posthog.models.filters.retention_filter import RetentionFilter
 from posthog.models.filters.stickiness_filter import StickinessFilter
 
 FilterType = Union[Filter, PathFilter, RetentionFilter, StickinessFilter]
-
-
-class AvailableFeature(str, Enum):
-    ZAPIER = "zapier"
-    ORGANIZATIONS_PROJECTS = "organizations_projects"
-    GOOGLE_LOGIN = "google_login"
-    SAML = "saml"
-    DASHBOARD_COLLABORATION = "dashboard_collaboration"
-    INGESTION_TAXONOMY = "ingestion_taxonomy"
